Checkbox update - all docks ok.

+++ include/ekg/draw/draw.hpp
@@ -5,7 +5,7 @@
 
 namespace ekg::draw {
     void rect(const ekg::rect &rect, const ekg::vec4 &color, int32_t line_thickness = 0);
-    void rect(int32_t x, int32_t y, int32_t w, int32_t h, const ekg::vec4 &color, int32_t line_thickness = 0);
+    void rect(float x, float y, float w, float h, const ekg::vec4 &color, int32_t line_thickness = 0);
     void scissor(int32_t x, int32_t y, int32_t w, int32_t h);
     void pop_scissor();
 }
